Move Tag related objects into their own module. Clean up fetching references
